requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.10.7
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
Markdown==2.6.8
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
PyJWT==1.5.0
pyparsing==2.2.0
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
svgwrite==1.1.11
Tree==0.1b8
urllib3==1.21.1
whitenoise==3.3.0
psycopg2==2.6.2
